Implemented python tests with pytest
