+++ tests/python/requirements.txt
@@ -1,2 +1,3 @@
+pytest
 psycopg2==2.9.3
 psutil==5.9.1
